Move UpdateJoinEvent to listeners package
